requirements.txt
bcrypt==4.1.1
falcon==3.1.3
greenlet==3.0.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
redis==5.0.1
SQLAlchemy==2.0.23
typing_extensions==4.9.0
